llvm/clang-5.0: fixes for Tiger
this version of llvm/clang-5.0 does actually build
and install on Tiger Intel, however the install names
and some of the library linkages are presently broken

because I have not yet figured out how to add the PROPER fix in the
cmake file to fix the @rpath links into full library path links

SO - I just did it all manually post-installation, and I have detailed how
to do it in the bottom of the Portfile.

It could either be done as a hack in the Portfile next time I get around to it
or maybe I wlll spend the time to sort out how to do it fully correctly
using cmake and finish a proper link procedure

anyway -- it works! clang-5.0 installs, and emulated_tls can then be
built in libcxx, and then thread_local storage works using clang-5.0
and libcxx on Tiger Intel.