meson 53: improve situaiton for Tiger
this patch stops the production of -rpath statements
that Tiger does not support. This works fairly well, but
can have issues if the port tries to build tools in the
build directory that use libraries that are also built
in the build directory. In this case, these tools will only
function if the tool happens to be in the same directory as
the library. You can copy the library manually over sometimes
to fix build issues (like babl). Not perfect, but what is
prerfect with TIger.

A better fix would be to build the dylibs with the full path
to their location in the build directory, and then rewrite that
with the full install path on installation.